+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - `UGGrid` index sets can now compute the indices of vertices of edges.
 - `UGGrid`: Fixed a bug in the numbering of prism edges.
+- Various bugs have been fixed in the `UGGrid` subdomain communication implementation.
 - Python bindings have been moved from the `dune-python` module which is now
   obsolete. To activate Python bindings the CMake flag
   `DUNE_ENABLE_PYTHONBINDINGS` needs to be turned on (default is off).